Create useClipboardWithNotification that shows an error or success message when copying to clipboard Consolidate request details section into DetailsContent component
