+++ src/pages/admin/gitpoap/requests/[id].tsx
@@ -1,41 +1,40 @@
-import { Center, Text } from '@mantine/core';
+import { Container, Title } from '@mantine/core';
 import { GetStaticPaths, GetStaticPropsContext } from 'next';
-import { SSRData } from 'next-urql';
+import { SSRData, withUrqlClient } from 'next-urql';
 import Head from 'next/head';
 import React from 'react';
 import { ConnectGitHub } from '../../../../components/admin/ConnectGitHub';
 import { SingleRequestPage } from '../../../../components/admin/requests/SingleRequestPage';
 import { GitPoapRequestDocument, GitPoapRequestQuery } from '../../../../graphql/generated-gql';
 import { useUser } from '../../../../hooks/useUser';
-import { createSSRUrqlClient } from '../../../../lib/urql';
+import { createSSRUrqlClient, urqlClientOptions } from '../../../../lib/urql';
 import { NextPageWithLayout } from '../../../_app';
 
 type PageProps = {
   urqlState: SSRData;
-  gitPOAPRequest: GitPoapRequestQuery['gitPOAPRequest'];
+  gitPOAPRequest: GitPoapRequestQuery['gitPOAPRequest'] | null;
 };
 
 const AdminGitPOAPRequest: NextPageWithLayout<PageProps> = ({ gitPOAPRequest }) => {
   const user = useUser();
-
-  if (!gitPOAPRequest) {
-    return (
-      <Center>
-        <Text>{'Request not found'}</Text>
-      </Center>
-    );
-  }
-
   return (
     <>
       <Head>
         <title>{'Requests | GitPOAP'}</title>
         <meta name="description" content="GitPOAP Admin" />
       </Head>
-      {user?.permissions.isStaff ? (
-        <SingleRequestPage gitPOAPRequest={gitPOAPRequest} />
+      {gitPOAPRequest ? (
+        user?.permissions.isStaff ? (
+          <Container>
+            <SingleRequestPage gitPOAPRequest={gitPOAPRequest} />
+          </Container>
+        ) : (
+          <ConnectGitHub />
+        )
       ) : (
-        <ConnectGitHub />
+        <Container mt={64}>
+          <Title>{'Request not found'}</Title>
+        </Container>
       )}
     </>
   );
@@ -46,9 +45,9 @@ const AdminGitPOAPRequest: NextPageWithLayout<PageProps> = ({ gitPOAPRequest }) 
  */
 export const getStaticProps = async (
   context: GetStaticPropsContext<{ id: string }>,
-): Promise<{ props: PageProps }> => {
+): Promise<{ props: PageProps; revalidate: number }> => {
   const { client, ssrCache } = createSSRUrqlClient();
-  const gitPOAPRequestId = context.params?.id as string;
+  const gitPOAPRequestId = parseInt(context.params?.id as string);
   const results = await client
     ?.query<GitPoapRequestQuery>(GitPoapRequestDocument, {
       gitPOAPRequestId,
@@ -59,16 +58,20 @@ export const getStaticProps = async (
     props: {
       urqlState: ssrCache.extractData(),
       /* coalesce to null if no data is returned -> nextJS doesn't like 'undefined' */
-      gitPOAPRequest: results?.data?.gitPOAPRequest,
+      gitPOAPRequest: results?.data?.gitPOAPRequest ?? null,
     },
+    revalidate: 1,
   };
 };
 
-export const getStaticPaths: GetStaticPaths<{ slug: string }> = async () => {
+export const getStaticPaths: GetStaticPaths = async () => {
   return {
-    paths: [], // indicates that no page needs be created at build time
+    paths: [],
     fallback: 'blocking', // indicates the type of fallback
   };
 };
 
-export default AdminGitPOAPRequest;
+export default withUrqlClient(
+  () => urqlClientOptions,
+  { ssr: false }, // Important so we don't wrap our component in getInitialProps
+)(AdminGitPOAPRequest);
